backend/requirements.txt
python-docx
pdfplumber
fastapi
uvicorn
aiofiles
